--- package.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "bootstrap": "^5.3.3",
-    "react-bootstrap": "^2.10.9",
-    "react-icons": "^5.5.0"
-  }
-}
