+++ webapp/package.json
@@ -17,11 +17,12 @@
     "lucide-react": "^0.477.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.9",
-    "react-icons": "^5.5.0",
     "react-bootstrap-icons": "^1.11.5",
     "react-custom-scrollbars-2": "^4.5.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^15.4.1",
+    "react-icons": "^5.5.0",
+    "react-router-dom": "^7.2.0",
     "react-scripts": "5.0.1",
     "react-scrollbars-custom": "^4.1.1",
     "react-simple-typewriter": "^5.0.1",
