+++ webapp/src/App.js
@@ -3,14 +3,12 @@ import CssBaseline from '@mui/material/CssBaseline';
 import Container from '@mui/material/Container';
 import Home from './components/home/Home'; 
 
+import router from './components/routers/AppRouter';
+import { RouterProvider } from 'react-router-dom';
+
 function App() {
   return (
-    <Container component="main" maxWidth="xs">
-      <CssBaseline />
-
-      {/* Mostramos directamente el componente Home */}
-      
-    </Container>
+    <RouterProvider router={router}/>
   );
 }
