+++ webapp/src/components/Login.js
@@ -4,7 +4,7 @@ import axios from 'axios';
 import { Container, Typography, TextField, Button, Snackbar } from '@mui/material';
 import { Typewriter } from "react-simple-typewriter";
 
-const Login = () => {
+export const Login = () => {
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
   const [message, setMessage] = useState('');
@@ -31,7 +31,10 @@ const Login = () => {
         setMessage(message.data.answer);
       }
       // Extract data from the response
-      const { createdAt: userCreatedAt } = response.data;
+      const { createdAt: userCreatedAt, token } = response.data;
+
+      // Store the token in the session storage
+      sessionStorage.setItem('token', token);
 
       setCreatedAt(userCreatedAt);
       setLoginSuccess(true);
@@ -92,5 +95,3 @@ const Login = () => {
     </Container>
   );
 };
-
-export default Login;
